+++ deltafi-api/lib/deltafi/api/v1/metrics/flow.rb
@@ -32,7 +32,7 @@ module Deltafi
           class << self
             def summary(params:)
               ingress = <<-QUERY
-              summarize(groupByTags(seriesByTag('name=~stats_counts.(bytes_in|files_in)', 'action=IngressAction'), "sum", "name", "ingressFlow"), '5y', 'sum')
+              summarize(groupByTags(seriesByTag('name=~stats_counts.(bytes_in|files_in)', 'action=ingress'), "sum", "name", "ingressFlow"), '5y', 'sum')
               QUERY
               egress = <<-QUERY
               summarize(groupByTags(seriesByTag('name=~stats_counts.(bytes_out|files_out|files_errored|files_filtered|files_dropped)'), "sum", "name", "ingressFlow"), '5y', 'sum')