+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Update the metric reports to use the `ingress` action tag
 
 ### Tech-Debt/Refactor
+- Optimize batch resume and replay operations
 
 ### Security