+++ CHANGELOG.md
@@ -3,7 +3,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased] - Next release 0.98.5
+## [Unreleased] - Next release 0.98.6
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Tech-Debt/Refactor
+
+### Security
+
+### Upgrade and Migration
+
+## [0.98.5] - 2022-10-03
 
 ### Added
 - New `cancel` mutation for DeltaFiles
@@ -672,7 +690,8 @@ No changes.  UI update only
 ### Security
 - Forced all projects to log4j 2.17.0 to avoid CVEs
 
-[Unreleased]: https://gitlab.com/systolic/deltafi/deltafi/-/compare/0.98.4...main
+[Unreleased]: https://gitlab.com/systolic/deltafi/deltafi/-/compare/0.98.5...main
+[0.98.4]: https://gitlab.com/systolic/deltafi/deltafi/-/compare/0.98.4...0.98.5
 [0.98.4]: https://gitlab.com/systolic/deltafi/deltafi/-/compare/0.98.3...0.98.4
 [0.98.3]: https://gitlab.com/systolic/deltafi/deltafi/-/compare/0.98.2...0.98.3
 [0.98.2]: https://gitlab.com/systolic/deltafi/deltafi/-/compare/0.98.1...0.98.2